+++ js/main.js
@@ -1,1 +1,7 @@
-import './data.js';
+import {fillingObject} from './data.js';
+
+const SIMILAR_DATA = 10;
+
+const objects = new Array(SIMILAR_DATA).fill(null).map(() => fillingObject());
+
+objects;
